\name{gCMAP-package}
\alias{gCMAP-package}
\alias{gCMAP}
\docType{package}
\title{
Tools for Connectivity Map-like analyses
}
\description{

  The gCMAP package provides a toolkit for comparing differential gene
  expression profiles through gene set enrichment analysis. Starting
  from normalized microarray or RNA-seq gene expression values (stored
  in lists of ExpressionSet and CountDataSet objects) the package
  performs differential expression analysis using the limma or DESeq
  packages. Supplying a simple list of gene identifiers, global
  differential expression profiles or data from complete experiments
  as input, users can use a unified set of several well-known gene set
  enrichment analysis methods to retrieve experiments with similar
  changes in gene expression. To take into account the directionality
  of gene expression changes, gCMAPQuery introduces the SignedGeneSet
  class, directly extending GeneSet from the GSEABase package.  To
  increase performance of large queries, multiple gene sets are stored
  as sparse incidence matrices within CMAPCollection eSets. gCMAP
  offers implementations of 1. Fisher's exact test (Fisher, J R Stat
  Soc, 1922) 2. The "connectivity map" method (Lamb et al, Science,
  2006) 3. Parametric and non-parametric t-statistic summaries (Jiang
  & Gentleman, Bioinformatics, 2007) and 4. Wilcoxon / Mann-Whitney
  rank sum statistics (Wilcoxon, Biometrics Bulletin, 1945) as well as
  wrappers for the 5. camera (Wu & Smyth, Nucleic Acid Res, 2012)
  6. mroast and romer (Wu et al, Bioinformatics, 2010) functions from
  the limma package and 7. wraps the gsea method from the mgsa package 
  (Bauer et al, NAR, 2010). All methods return CMAPResult objects, an S4
  class inheriting from AnnotatedDataFrame, containing enrichment
  statistics as well as annotation data and providing simple
  high-level summary plots.
}
\details{
\tabular{ll}{
Package: \tab gCMAP\cr
Type: \tab Package\cr
Version: \tab 1.3.5\cr
Date: \tab 2013-03-11\cr
Depends: \tab GSEABase, limma (>= 3.15.14)\cr
Imports: \tab Biobase, BiocGenerics, methods, GSEAlm, Category, bigmemory, bigmemoryExtras (>=
1.1.2), Matrix (>= 1.0.9), parallel, annotate, genefilter, AnnotationDbi\cr
Suggests: \tab DESeq, KEGG.db, reactome.db, RUnit, GO.db, mgsa\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
OS_type: \tab unix\cr
ByteCompile: \tab TRUE\cr
biocViews: \tab Bioinformatics, Microarray, Software, Pathways, Annotation\cr
Collate: \tab 'AllClasses.R' 'AllGenerics.R' 'SignedGeneSet-accessors.R'
'utility-functions.R' 'camera_score-methods.R'
'connectivity_score-methods.R' 'featureScore-methods.R'
'fisher_score-methods.R' 'geneIndex-methods.R'
'gsealm_jg_score-methods.R' 'gsealm_score-methods.R'
'incidence-methods.R' 'mgsa_score-methods.R'
'mapIdentifiers-methods.R'
'minSetSize-methods.R' 'mroast_score-methods.R'
'romer_score-methods.R' 'wilcox_score-methods.R'
'CMAPCollection-accessors.R' 'CMAPResults-accessors.R'\cr
Built: \tab R 3.0.0; ; 2013-03-11 17:13:45 UTC; unix\cr
}

Index:
\preformatted{
.DESeq_nbinom           Function to perform a DESeq analysis to detect
                        differential expression between perturbation
                        and control groups.
CMAPCollection-class    Class '"CMAPCollection"'
CMAPResults-class       Class '"CMAPResults"'
GeneSet                 Methods for 'GeneSet' and 'GeneColorSet'
KEGG2cmap               Functions to generate species-specific
                        CMAPCollections from Bioconductor KEGG.db,
                        reactome.db or GO.db annotation packages or the
                        wikipathways <URL:
                        http://www.wikipathways.org/index.php/Download_Pathways>
                        project.
SignedGeneSet           Construtor for SignedGeneSet
SignedGeneSet-class     Class '"SignedGeneSet"'
annotate_eset_list      Function to complile a data frame with
                        per-instance annotation for a list of eSet
                        objects generated by the 'splitPerturbations'
                        function. The output can be used directly as
                        sample.annotation for the 'NChannelSet'
                        function.
camera_score-methods    Methods for Function 'camera_score' in Package
                        'gCMAP'
center_eSet             A function to to center columns of eSet
                        channels on either their kernel density peak,
                        their mean or their median.
connectivity_score      Broad CMAP gene set enrichment metrics
eSetOnDisk              A function to store the assayData of an eSet
                        object as BigMatrix files on disk.
eset_instances          A function to subset an eSet with expression
                        data into smaller datasets, each corresponding
                        to a single perturbation experiment.
featureScores-methods   Methods to obtain scores for CMAPCollection
                        gene sets from a matrix or eSet
fisher_score-methods    Hypergeometric probability of gene set
                        enrichment
gCMAP-package           Tools for Connectivity Map-like analyses
gCMAPData               Example 'NChannelSet'
geneIndex-methods       Methods for Function 'geneIndex' in Package
                        'gCMAP'
generate_gCMAP_NChannelSet
                        Generate a perturbation profile library from
                        expression sets of control/treatment pairs
gsealm_jg_score-methods
                        Parametric test for testing normally
                        distributed scores for gene set enrichment
gsealm_score-methods    Methods for Function 'gsealm_score' in Package
                        'gCMAP'
induceCMAPCollection-methods
                        Methods for Function 'induceCMAPCollection' in
                        Package 'gCMAP'
mapNmerge               A function to map eSet featureNames and
                        calculate summaries for many-to-one mapping
                        features
matrix_or_big.matrix-class
                        Class '"matrix_or_big.matrix"'
memorize                Create a new NChannelSet instance by selecting
                        specific channels and load BigMatrix assayData
                        into memory, if present
mergeCMAPs              This function merged two eSets.
mgsa_score-methods      Model-based gene set analysis (MGSA)
minSetSize              GeneSetCollection length filtering
mroast_score-methods    Methods for Function 'mroast_score' in Package
                        'gCMAP'
pairwise_DESeq          Generate statistics associated with pairwise
                        differential expression from RNAseq count data
pairwise_compare        Generate statistics associated with pairwise
                        differential expression
romer_score-methods     Methods for Function 'romer_score' in Package
                        'gCMAP'
signedRankSumTest       An implementation of the Wilcox rank sum test /
                        Mann-Whitney test that takes into account the
                        direction / sign of gene set members and
                        possibly the correlation between cases
splitPerturbations      Function to split an ExpressionSet downloaded
                        from ArrayExpress based on the experimental
                        factors present in the phenoData slot
wilcox_score-methods    Methods for Function 'wilcox_score' in Package
                        'gCMAP'
zScores                 Function to calculate z-scores from p-values
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{diffExprAnalysis} \tab main (source)\cr
\code{gCMAP} \tab main (source)\cr
}
}
\author{
Thomas Sandmann <sandmann.thomas@gene.com>, Richard Bourgon
	<bourgon.richard@gene.com> and Sarah Kummerfeld
	<kummerfeld.sarah@gene.com>

Maintainer: Thomas Sandmann <sandmann.thomas@gene.com>
}
\keyword{ package }

