\name{GeneSet}

\docType{methods}

\alias{GeneSet}
\alias{GeneSet,NChannelSet-method}
\alias{GeneColorSet,NChannelSet-method}

\title{Methods for \code{GeneSet} and \code{GeneColorSet}}

\description{
 Additional methods for function \code{GeneSet} and \code{GeneColorSet},
 supporting use of \code{\linkS4class{NChannelSet}} templates.
}

\note{ The \code{ExpressionSet} methods are used verbatim, since only
  metadata is utilized. Note that \code{collectionType} will be
  \code{ExpressionSet} as a result. }

\keyword{methods}
